Jpk/account select

+++ src/Payments.Mvc/Controllers/InvoicesController.cs
@@ -63,9 +63,26 @@ namespace Payments.Mvc.Controllers
         public async Task<IActionResult> Create()
         {
             var team = await _dbContext.Teams
+                .Include(t => t.Accounts)
                 .FirstOrDefaultAsync(t => t.Slug == TeamSlug);
 
             ViewBag.Team = new { team.Id, team.Name, team.Slug };
+
+            ViewBag.Accounts = team.Accounts
+                .Where(a => a.IsActive)
+                .Select(a => new
+                {
+                    a.Id,
+                    a.Name,
+                    a.Description,
+                    a.IsDefault,
+                    a.Chart,
+                    a.Account,
+                    a.SubAccount,
+                    a.Object,
+                    a.SubObject,
+                    a.Project,
+                });
 
             return View();
         }
@@ -75,6 +92,7 @@ namespace Payments.Mvc.Controllers
         {
             // look for invoice
             var invoice = await _dbContext.Invoices
+                .Include(i => i.Account)
                 .Include(i => i.Items)
                 .Where(i => i.Team.Slug == TeamSlug)
                 .FirstOrDefaultAsync(i => i.Id == id);
@@ -86,13 +104,31 @@ namespace Payments.Mvc.Controllers
 
             // fetch team data
             var team = await _dbContext.Teams
+                .Include(t => t.Accounts)
                 .FirstOrDefaultAsync(t => t.Slug == TeamSlug);
 
             ViewBag.Team = new { team.Id, team.Name, team.Slug };
 
+            ViewBag.Accounts = team.Accounts
+                .Where(a => a.IsActive)
+                .Select(a => new
+                {
+                    a.Id,
+                    a.Name,
+                    a.Description,
+                    a.IsDefault,
+                    a.Chart,
+                    a.Account,
+                    a.SubAccount,
+                    a.Object,
+                    a.SubObject,
+                    a.Project,
+                });
+
             // build model for view
             var model = new EditInvoiceViewModel()
             {
+                AccountId = invoice.Account.Id,
                 Discount = invoice.Discount,
                 Tax = invoice.TaxPercent,
                 Memo = invoice.Memo,
@@ -123,6 +159,30 @@ namespace Payments.Mvc.Controllers
             var user = await _userManager.GetUserAsync(User);
             var team = await _dbContext.Teams.FirstOrDefaultAsync(t => t.Slug == TeamSlug);
 
+            // find account
+            var account = await _dbContext.FinancialAccounts
+                .FirstOrDefaultAsync(a => a.Team.Slug == TeamSlug && a.Id == model.AccountId);
+
+            if (account == null)
+            {
+                ModelState.AddModelError("AccountId", "Account Id not found for this team.");
+            }
+            else if (!account.IsActive)
+            {
+                ModelState.AddModelError("AccountId", "Account is inactive.");
+            }
+
+            // validate model
+            if (!ModelState.IsValid)
+            {
+                return new JsonResult(new
+                {
+                    success = false,
+                    errorMessage = "Errors found in request",
+                    modelState = ModelState
+                });
+            }
+
             // manage multiple customer scenario
             var invoices = new List<Invoice>();
             foreach (var customer in model.Customers)
@@ -130,6 +190,7 @@ namespace Payments.Mvc.Controllers
                 // create new object, track it
                 var invoice = new Invoice
                 {
+                    Account         = account,
                     Creator         = user,
                     Team            = team,
                     Discount        = model.Discount,
@@ -182,10 +243,35 @@ namespace Payments.Mvc.Controllers
                 return NotFound();
             }
 
+            // find account
+            var account = await _dbContext.FinancialAccounts
+                .FirstOrDefaultAsync(a => a.Team.Slug == TeamSlug && a.Id == model.AccountId);
+
+            if (account == null)
+            {
+                ModelState.AddModelError("AccountId", "Account Id not found for this team.");
+            }
+            else if (!account.IsActive)
+            {
+                ModelState.AddModelError("AccountId", "Account is inactive.");
+            }
+
+            // validate model
+            if (!ModelState.IsValid)
+            {
+                return new JsonResult(new
+                {
+                    success = false,
+                    errorMessage = "Errors found in request",
+                    modelState = ModelState
+                });
+            }
+
             // remove old items
             _dbContext.LineItems.RemoveRange(invoice.Items);
 
             // update invoice
+            invoice.Account         = account;
             invoice.CustomerAddress = model.Customer.Address;
             invoice.CustomerEmail   = model.Customer.Email;
             invoice.CustomerName    = model.Customer.Name;
@@ -266,7 +352,6 @@ namespace Payments.Mvc.Controllers
         {
             // find item
             var invoice = await _dbContext.Invoices
-                .Include(i => i.Items)
                 .Where(i => i.Team.Slug == TeamSlug)
                 .FirstOrDefaultAsync(i => i.Id == id);
